Migrate more custom matrix lite code to core
