+++ keyboards/xd84/rules.mk
@@ -37,7 +37,7 @@ LINK_TIME_OPTIMIZATION_ENABLE = yes
 CUSTOM_MATRIX = lite
 
 VPATH += drivers/gpio
-SRC += custom_matrix_helper.c pca9555.c matrix.c
+SRC += pca9555.c matrix.c
 QUANTUM_LIB_SRC += i2c_master.c
 
 LAYOUTS = 75_ansi 75_iso
